Branchement du wizard preinscription avec Backend.

+++ src/main/java/com/gsr/myschool/server/business/EtablissementScolaire.java
@@ -6,17 +6,17 @@ import javax.persistence.*;
 public class EtablissementScolaire implements java.io.Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @Column(unique = true)
     private String nom;
     private Boolean reference;
     private String adresse;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer newId) {
+    public void setId(Long newId) {
         this.id = newId;
     }
